Refactor card code
