+++ custom_components/lock_code_manager/www/lock-code-manager-strategy.js
@@ -1,1 +1,1 @@
-var t,e,n,r,i,o;function a(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function l(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?a(Object(n),!0).forEach((function(e){u(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):a(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function u(t,e,n){return(e=P(e))in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}function s(){return s="undefined"!=typeof Reflect&&Reflect.get?Reflect.get.bind():function(t,e,n){var r=function(t,e){for(;!Object.prototype.hasOwnProperty.call(t,e)&&null!==(t=w(t)););return t}(t,e);if(r){var i=Object.getOwnPropertyDescriptor(r,e);return i.get?i.get.call(arguments.length<3?t:n):i.value}},s.apply(this,arguments)}function c(t){return c="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t},c(t)}function h(t){return function(t){if(Array.isArray(t))return E(t)}(t)||function(t){if("undefined"!=typeof Symbol&&null!=t[Symbol.iterator]||null!=t["@@iterator"])return Array.from(t)}(t)||A(t)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function f(t,e){return function(t){if(Array.isArray(t))return t}(t)||function(t,e){var n=null==t?null:"undefined"!=typeof Symbol&&t[Symbol.iterator]||t["@@iterator"];if(null!=n){var r,i,o,a,l=[],u=!0,s=!1;try{if(o=(n=n.call(t)).next,0===e){if(Object(n)!==n)return;u=!1}else for(;!(u=(r=o.call(n)).done)&&(l.push(r.value),l.length!==e);u=!0);}catch(t){s=!0,i=t}finally{try{if(!u&&null!=n.return&&(a=n.return(),Object(a)!==a))return}finally{if(s)throw i}}return l}}(t,e)||A(t,e)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function d(){d=function(){return e};var t,e={},n=Object.prototype,r=n.hasOwnProperty,i=Object.defineProperty||function(t,e,n){t[e]=n.value},o="function"==typeof Symbol?Symbol:{},a=o.iterator||"@@iterator",l=o.asyncIterator||"@@asyncIterator",u=o.toStringTag||"@@toStringTag";function s(t,e,n){return Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}),t[e]}try{s({},"")}catch(t){s=function(t,e,n){return t[e]=n}}function h(t,e,n,r){var o=e&&e.prototype instanceof _?e:_,a=Object.create(o.prototype),l=new N(r||[]);return i(a,"_invoke",{value:x(t,n,l)}),a}function f(t,e,n){try{return{type:"normal",arg:t.call(e,n)}}catch(t){return{type:"throw",arg:t}}}e.wrap=h;var v="suspendedStart",p="suspendedYield",y="executing",g="completed",m={};function _(){}function b(){}function $(){}var w={};s(w,a,(function(){return this}));var k=Object.getPrototypeOf,A=k&&k(k(L([])));A&&A!==n&&r.call(A,a)&&(w=A);var E=$.prototype=_.prototype=Object.create(w);function S(t){["next","throw","return"].forEach((function(e){s(t,e,(function(t){return this._invoke(e,t)}))}))}function O(t,e){function n(i,o,a,l){var u=f(t[i],t,o);if("throw"!==u.type){var s=u.arg,h=s.value;return h&&"object"==c(h)&&r.call(h,"__await")?e.resolve(h.__await).then((function(t){n("next",t,a,l)}),(function(t){n("throw",t,a,l)})):e.resolve(h).then((function(t){s.value=t,a(s)}),(function(t){return n("throw",t,a,l)}))}l(u.arg)}var o;i(this,"_invoke",{value:function(t,r){function i(){return new e((function(e,i){n(t,r,e,i)}))}return o=o?o.then(i,i):i()}})}function x(e,n,r){var i=v;return function(o,a){if(i===y)throw new Error("Generator is already running");if(i===g){if("throw"===o)throw a;return{value:t,done:!0}}for(r.method=o,r.arg=a;;){var l=r.delegate;if(l){var u=P(l,r);if(u){if(u===m)continue;return u}}if("next"===r.method)r.sent=r._sent=r.arg;else if("throw"===r.method){if(i===v)throw i=g,r.arg;r.dispatchException(r.arg)}else"return"===r.method&&r.abrupt("return",r.arg);i=y;var s=f(e,n,r);if("normal"===s.type){if(i=r.done?g:p,s.arg===m)continue;return{value:s.arg,done:r.done}}"throw"===s.type&&(i=g,r.method="throw",r.arg=s.arg)}}}function P(e,n){var r=n.method,i=e.iterator[r];if(i===t)return n.delegate=null,"throw"===r&&e.iterator.return&&(n.method="return",n.arg=t,P(e,n),"throw"===n.method)||"return"!==r&&(n.method="throw",n.arg=new TypeError("The iterator does not provide a '"+r+"' method")),m;var o=f(i,e.iterator,n.arg);if("throw"===o.type)return n.method="throw",n.arg=o.arg,n.delegate=null,m;var a=o.arg;return a?a.done?(n[e.resultName]=a.value,n.next=e.nextLoc,"return"!==n.method&&(n.method="next",n.arg=t),n.delegate=null,m):a:(n.method="throw",n.arg=new TypeError("iterator result is not an object"),n.delegate=null,m)}function j(t){var e={tryLoc:t[0]};1 in t&&(e.catchLoc=t[1]),2 in t&&(e.finallyLoc=t[2],e.afterLoc=t[3]),this.tryEntries.push(e)}function C(t){var e=t.completion||{};e.type="normal",delete e.arg,t.completion=e}function N(t){this.tryEntries=[{tryLoc:"root"}],t.forEach(j,this),this.reset(!0)}function L(e){if(e||""===e){var n=e[a];if(n)return n.call(e);if("function"==typeof e.next)return e;if(!isNaN(e.length)){var i=-1,o=function n(){for(;++i<e.length;)if(r.call(e,i))return n.value=e[i],n.done=!1,n;return n.value=t,n.done=!0,n};return o.next=o}}throw new TypeError(c(e)+" is not iterable")}return b.prototype=$,i(E,"constructor",{value:$,configurable:!0}),i($,"constructor",{value:b,configurable:!0}),b.displayName=s($,u,"GeneratorFunction"),e.isGeneratorFunction=function(t){var e="function"==typeof t&&t.constructor;return!!e&&(e===b||"GeneratorFunction"===(e.displayName||e.name))},e.mark=function(t){return Object.setPrototypeOf?Object.setPrototypeOf(t,$):(t.__proto__=$,s(t,u,"GeneratorFunction")),t.prototype=Object.create(E),t},e.awrap=function(t){return{__await:t}},S(O.prototype),s(O.prototype,l,(function(){return this})),e.AsyncIterator=O,e.async=function(t,n,r,i,o){void 0===o&&(o=Promise);var a=new O(h(t,n,r,i),o);return e.isGeneratorFunction(n)?a:a.next().then((function(t){return t.done?t.value:a.next()}))},S(E),s(E,u,"Generator"),s(E,a,(function(){return this})),s(E,"toString",(function(){return"[object Generator]"})),e.keys=function(t){var e=Object(t),n=[];for(var r in e)n.push(r);return n.reverse(),function t(){for(;n.length;){var r=n.pop();if(r in e)return t.value=r,t.done=!1,t}return t.done=!0,t}},e.values=L,N.prototype={constructor:N,reset:function(e){if(this.prev=0,this.next=0,this.sent=this._sent=t,this.done=!1,this.delegate=null,this.method="next",this.arg=t,this.tryEntries.forEach(C),!e)for(var n in this)"t"===n.charAt(0)&&r.call(this,n)&&!isNaN(+n.slice(1))&&(this[n]=t)},stop:function(){this.done=!0;var t=this.tryEntries[0].completion;if("throw"===t.type)throw t.arg;return this.rval},dispatchException:function(e){if(this.done)throw e;var n=this;function i(r,i){return l.type="throw",l.arg=e,n.next=r,i&&(n.method="next",n.arg=t),!!i}for(var o=this.tryEntries.length-1;o>=0;--o){var a=this.tryEntries[o],l=a.completion;if("root"===a.tryLoc)return i("end");if(a.tryLoc<=this.prev){var u=r.call(a,"catchLoc"),s=r.call(a,"finallyLoc");if(u&&s){if(this.prev<a.catchLoc)return i(a.catchLoc,!0);if(this.prev<a.finallyLoc)return i(a.finallyLoc)}else if(u){if(this.prev<a.catchLoc)return i(a.catchLoc,!0)}else{if(!s)throw new Error("try statement without catch or finally");if(this.prev<a.finallyLoc)return i(a.finallyLoc)}}}},abrupt:function(t,e){for(var n=this.tryEntries.length-1;n>=0;--n){var i=this.tryEntries[n];if(i.tryLoc<=this.prev&&r.call(i,"finallyLoc")&&this.prev<i.finallyLoc){var o=i;break}}o&&("break"===t||"continue"===t)&&o.tryLoc<=e&&e<=o.finallyLoc&&(o=null);var a=o?o.completion:{};return a.type=t,a.arg=e,o?(this.method="next",this.next=o.finallyLoc,m):this.complete(a)},complete:function(t,e){if("throw"===t.type)throw t.arg;return"break"===t.type||"continue"===t.type?this.next=t.arg:"return"===t.type?(this.rval=this.arg=t.arg,this.method="return",this.next="end"):"normal"===t.type&&e&&(this.next=e),m},finish:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var n=this.tryEntries[e];if(n.finallyLoc===t)return this.complete(n.completion,n.afterLoc),C(n),m}},catch:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var n=this.tryEntries[e];if(n.tryLoc===t){var r=n.completion;if("throw"===r.type){var i=r.arg;C(n)}return i}}throw new Error("illegal catch attempt")},delegateYield:function(e,n,r){return this.delegate={iterator:L(e),resultName:n,nextLoc:r},"next"===this.method&&(this.arg=t),m}},e}function v(t,e,n,r,i,o,a){try{var l=t[o](a),u=l.value}catch(t){return void n(t)}l.done?e(u):Promise.resolve(u).then(r,i)}function p(t){return function(){var e=this,n=arguments;return new Promise((function(r,i){var o=t.apply(e,n);function a(t){v(o,r,i,a,l,"next",t)}function l(t){v(o,r,i,a,l,"throw",t)}a(void 0)}))}}function y(t,e,n){return e=w(e),function(t,e){if(e&&("object"===c(e)||"function"==typeof e))return e;if(void 0!==e)throw new TypeError("Derived constructors may only return object or undefined");return g(t)}(t,b()?Reflect.construct(e,n||[],w(t).constructor):e.apply(t,n))}function g(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}function m(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Super expression must either be null or a function");t.prototype=Object.create(e&&e.prototype,{constructor:{value:t,writable:!0,configurable:!0}}),Object.defineProperty(t,"prototype",{writable:!1}),e&&$(t,e)}function _(t){var e="function"==typeof Map?new Map:void 0;return _=function(t){if(null===t||!function(t){try{return-1!==Function.toString.call(t).indexOf("[native code]")}catch(e){return"function"==typeof t}}(t))return t;if("function"!=typeof t)throw new TypeError("Super expression must either be null or a function");if(void 0!==e){if(e.has(t))return e.get(t);e.set(t,n)}function n(){return function(t,e,n){if(b())return Reflect.construct.apply(null,arguments);var r=[null];r.push.apply(r,e);var i=new(t.bind.apply(t,r));return n&&$(i,n.prototype),i}(t,arguments,w(this).constructor)}return n.prototype=Object.create(t.prototype,{constructor:{value:n,enumerable:!1,writable:!0,configurable:!0}}),$(n,t)},_(t)}function b(){try{var t=!Boolean.prototype.valueOf.call(Reflect.construct(Boolean,[],(function(){})))}catch(t){}return(b=function(){return!!t})()}function $(t,e){return $=Object.setPrototypeOf?Object.setPrototypeOf.bind():function(t,e){return t.__proto__=e,t},$(t,e)}function w(t){return w=Object.setPrototypeOf?Object.getPrototypeOf.bind():function(t){return t.__proto__||Object.getPrototypeOf(t)},w(t)}function k(t,e){var n="undefined"!=typeof Symbol&&t[Symbol.iterator]||t["@@iterator"];if(!n){if(Array.isArray(t)||(n=A(t))||e&&t&&"number"==typeof t.length){n&&(t=n);var r=0,i=function(){};return{s:i,n:function(){return r>=t.length?{done:!0}:{done:!1,value:t[r++]}},e:function(t){throw t},f:i}}throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}var o,a=!0,l=!1;return{s:function(){n=n.call(t)},n:function(){var t=n.next();return a=t.done,t},e:function(t){l=!0,o=t},f:function(){try{a||null==n.return||n.return()}finally{if(l)throw o}}}}function A(t,e){if(t){if("string"==typeof t)return E(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);return"Object"===n&&t.constructor&&(n=t.constructor.name),"Map"===n||"Set"===n?Array.from(t):"Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)?E(t,e):void 0}}function E(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}function S(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function O(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,P(r.key),r)}}function x(t,e,n){return e&&O(t.prototype,e),n&&O(t,n),Object.defineProperty(t,"prototype",{writable:!1}),t}function P(t){var e=function(t,e){if("object"!=c(t)||!t)return t;var n=t[Symbol.toPrimitive];if(void 0!==n){var r=n.call(t,e||"default");if("object"!=c(r))return r;throw new TypeError("@@toPrimitive must return a primitive value.")}return("string"===e?String:Number)(t)}(t,"string");return"symbol"==c(e)?e:String(e)}var j=globalThis,C=j.ShadowRoot&&(void 0===j.ShadyCSS||j.ShadyCSS.nativeShadow)&&"adoptedStyleSheets"in Document.prototype&&"replace"in CSSStyleSheet.prototype,N=Symbol(),L=new WeakMap,T=function(){function t(e,n,r){if(S(this,t),this._$cssResult$=!0,r!==N)throw Error("CSSResult is not constructable. Use `unsafeCSS` or `css` instead.");this.cssText=e,this.t=n}return x(t,[{key:"styleSheet",get:function(){var t=this.o,e=this.t;if(C&&void 0===t){var n=void 0!==e&&1===e.length;n&&(t=L.get(e)),void 0===t&&((this.o=t=new CSSStyleSheet).replaceSync(this.cssText),n&&L.set(e,t))}return t}},{key:"toString",value:function(){return this.cssText}}]),t}(),R=C?function(t){return t}:function(t){return t instanceof CSSStyleSheet?function(t){var e,n="",r=k(t.cssRules);try{for(r.s();!(e=r.n()).done;){n+=e.value.cssText}}catch(t){r.e(t)}finally{r.f()}return function(t){return new T("string"==typeof t?t:t+"",void 0,N)}(n)}(t):t},U=Object.is,I=Object.defineProperty,M=Object.getOwnPropertyDescriptor,H=Object.getOwnPropertyNames,D=Object.getOwnPropertySymbols,z=Object.getPrototypeOf,B=globalThis,W=B.trustedTypes,V=W?W.emptyScript:"",G=B.reactiveElementPolyfillSupport,F=function(t,e){return t},q={toAttribute:function(t,e){switch(e){case Boolean:t=t?V:null;break;case Object:case Array:t=null==t?t:JSON.stringify(t)}return t},fromAttribute:function(t,e){var n=t;switch(e){case Boolean:n=null!==t;break;case Number:n=null===t?null:Number(t);break;case Object:case Array:try{n=JSON.parse(t)}catch(t){n=null}}return n}},J=function(t,e){return!U(t,e)},K={attribute:!0,type:String,converter:q,reflect:!1,hasChanged:J};null!==(t=Symbol.metadata)&&void 0!==t||(Symbol.metadata=Symbol("metadata")),null!==(e=B.litPropertyMetadata)&&void 0!==e||(B.litPropertyMetadata=new WeakMap);var Y=function(t){function e(){var t;return S(this,e),(t=y(this,e))._$Ep=void 0,t.isUpdatePending=!1,t.hasUpdated=!1,t._$Em=null,t._$Ev(),t}var n;return m(e,_(HTMLElement)),x(e,[{key:"_$Ev",value:function(){var t,e=this;this._$ES=new Promise((function(t){return e.enableUpdating=t})),this._$AL=new Map,this._$E_(),this.requestUpdate(),null===(t=this.constructor.l)||void 0===t||t.forEach((function(t){return t(e)}))}},{key:"addController",value:function(t){var e,n;(null!==(e=this._$EO)&&void 0!==e?e:this._$EO=new Set).add(t),void 0!==this.renderRoot&&this.isConnected&&(null===(n=t.hostConnected)||void 0===n||n.call(t))}},{key:"removeController",value:function(t){var e;null===(e=this._$EO)||void 0===e||e.delete(t)}},{key:"_$E_",value:function(){var t,e=new Map,n=k(this.constructor.elementProperties.keys());try{for(n.s();!(t=n.n()).done;){var r=t.value;this.hasOwnProperty(r)&&(e.set(r,this[r]),delete this[r])}}catch(t){n.e(t)}finally{n.f()}e.size>0&&(this._$Ep=e)}},{key:"createRenderRoot",value:function(){var t,e=null!==(t=this.shadowRoot)&&void 0!==t?t:this.attachShadow(this.constructor.shadowRootOptions);return function(t,e){if(C)t.adoptedStyleSheets=e.map((function(t){return t instanceof CSSStyleSheet?t:t.styleSheet}));else{var n,r=k(e);try{for(r.s();!(n=r.n()).done;){var i=n.value,o=document.createElement("style"),a=j.litNonce;void 0!==a&&o.setAttribute("nonce",a),o.textContent=i.cssText,t.appendChild(o)}}catch(t){r.e(t)}finally{r.f()}}}(e,this.constructor.elementStyles),e}},{key:"connectedCallback",value:function(){var t,e;null!==(t=this.renderRoot)&&void 0!==t||(this.renderRoot=this.createRenderRoot()),this.enableUpdating(!0),null===(e=this._$EO)||void 0===e||e.forEach((function(t){var e;return null===(e=t.hostConnected)||void 0===e?void 0:e.call(t)}))}},{key:"enableUpdating",value:function(t){}},{key:"disconnectedCallback",value:function(){var t;null===(t=this._$EO)||void 0===t||t.forEach((function(t){var e;return null===(e=t.hostDisconnected)||void 0===e?void 0:e.call(t)}))}},{key:"attributeChangedCallback",value:function(t,e,n){this._$AK(t,n)}},{key:"_$EC",value:function(t,e){var n=this.constructor.elementProperties.get(t),r=this.constructor._$Eu(t,n);if(void 0!==r&&!0===n.reflect){var i,o=(void 0!==(null===(i=n.converter)||void 0===i?void 0:i.toAttribute)?n.converter:q).toAttribute(e,n.type);this._$Em=t,null==o?this.removeAttribute(r):this.setAttribute(r,o),this._$Em=null}}},{key:"_$AK",value:function(t,e){var n=this.constructor,r=n._$Eh.get(t);if(void 0!==r&&this._$Em!==r){var i,o=n.getPropertyOptions(r),a="function"==typeof o.converter?{fromAttribute:o.converter}:void 0!==(null===(i=o.converter)||void 0===i?void 0:i.fromAttribute)?o.converter:q;this._$Em=r,this[r]=a.fromAttribute(e,o.type),this._$Em=null}}},{key:"requestUpdate",value:function(t,e,n){if(void 0!==t){var r,i;if(null!==(r=n)&&void 0!==r||(n=this.constructor.getPropertyOptions(t)),!(null!==(i=n.hasChanged)&&void 0!==i?i:J)(this[t],e))return;this.P(t,e,n)}!1===this.isUpdatePending&&(this._$ES=this._$ET())}},{key:"P",value:function(t,e,n){var r;this._$AL.has(t)||this._$AL.set(t,e),!0===n.reflect&&this._$Em!==t&&(null!==(r=this._$Ej)&&void 0!==r?r:this._$Ej=new Set).add(t)}},{key:"_$ET",value:(n=p(d().mark((function t(){var e;return d().wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return this.isUpdatePending=!0,t.prev=1,t.next=4,this._$ES;case 4:t.next=9;break;case 6:t.prev=6,t.t0=t.catch(1),Promise.reject(t.t0);case 9:if(e=this.scheduleUpdate(),t.t1=null!=e,!t.t1){t.next=14;break}return t.next=14,e;case 14:return t.abrupt("return",!this.isUpdatePending);case 15:case"end":return t.stop()}}),t,this,[[1,6]])}))),function(){return n.apply(this,arguments)})},{key:"scheduleUpdate",value:function(){return this.performUpdate()}},{key:"performUpdate",value:function(){if(this.isUpdatePending){if(!this.hasUpdated){var t;if(null!==(t=this.renderRoot)&&void 0!==t||(this.renderRoot=this.createRenderRoot()),this._$Ep){var e,n=k(this._$Ep);try{for(n.s();!(e=n.n()).done;){var r=f(e.value,2),i=r[0],o=r[1];this[i]=o}}catch(t){n.e(t)}finally{n.f()}this._$Ep=void 0}var a=this.constructor.elementProperties;if(a.size>0){var l,u=k(a);try{for(u.s();!(l=u.n()).done;){var s=f(l.value,2),c=s[0],h=s[1];!0!==h.wrapped||this._$AL.has(c)||void 0===this[c]||this.P(c,this[c],h)}}catch(t){u.e(t)}finally{u.f()}}}var d=!1,v=this._$AL;try{var p;(d=this.shouldUpdate(v))?(this.willUpdate(v),null!==(p=this._$EO)&&void 0!==p&&p.forEach((function(t){var e;return null===(e=t.hostUpdate)||void 0===e?void 0:e.call(t)})),this.update(v)):this._$EU()}catch(v){throw d=!1,this._$EU(),v}d&&this._$AE(v)}}},{key:"willUpdate",value:function(t){}},{key:"_$AE",value:function(t){var e;null!==(e=this._$EO)&&void 0!==e&&e.forEach((function(t){var e;return null===(e=t.hostUpdated)||void 0===e?void 0:e.call(t)})),this.hasUpdated||(this.hasUpdated=!0,this.firstUpdated(t)),this.updated(t)}},{key:"_$EU",value:function(){this._$AL=new Map,this.isUpdatePending=!1}},{key:"updateComplete",get:function(){return this.getUpdateComplete()}},{key:"getUpdateComplete",value:function(){return this._$ES}},{key:"shouldUpdate",value:function(t){return!0}},{key:"update",value:function(t){var e=this;this._$Ej&&(this._$Ej=this._$Ej.forEach((function(t){return e._$EC(t,e[t])}))),this._$EU()}},{key:"updated",value:function(t){}},{key:"firstUpdated",value:function(t){}}],[{key:"addInitializer",value:function(t){var e;this._$Ei(),(null!==(e=this.l)&&void 0!==e?e:this.l=[]).push(t)}},{key:"observedAttributes",get:function(){return this.finalize(),this._$Eh&&h(this._$Eh.keys())}},{key:"createProperty",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:K;if(e.state&&(e.attribute=!1),this._$Ei(),this.elementProperties.set(t,e),!e.noAccessor){var n=Symbol(),r=this.getPropertyDescriptor(t,n,e);void 0!==r&&I(this.prototype,t,r)}}},{key:"getPropertyDescriptor",value:function(t,e,n){var r,i=null!==(r=M(this.prototype,t))&&void 0!==r?r:{get:function(){return this[e]},set:function(t){this[e]=t}},o=i.get,a=i.set;return{get:function(){return null==o?void 0:o.call(this)},set:function(e){var r=null==o?void 0:o.call(this);a.call(this,e),this.requestUpdate(t,r,n)},configurable:!0,enumerable:!0}}},{key:"getPropertyOptions",value:function(t){var e;return null!==(e=this.elementProperties.get(t))&&void 0!==e?e:K}},{key:"_$Ei",value:function(){if(!this.hasOwnProperty(F("elementProperties"))){var t=z(this);t.finalize(),void 0!==t.l&&(this.l=h(t.l)),this.elementProperties=new Map(t.elementProperties)}}},{key:"finalize",value:function(){if(!this.hasOwnProperty(F("finalized"))){if(this.finalized=!0,this._$Ei(),this.hasOwnProperty(F("properties"))){var t,e=this.properties,n=k([].concat(h(H(e)),h(D(e))));try{for(n.s();!(t=n.n()).done;){var r=t.value;this.createProperty(r,e[r])}}catch(t){n.e(t)}finally{n.f()}}var i=this[Symbol.metadata];if(null!==i){var o=litPropertyMetadata.get(i);if(void 0!==o){var a,l=k(o);try{for(l.s();!(a=l.n()).done;){var u=f(a.value,2),s=u[0],c=u[1];this.elementProperties.set(s,c)}}catch(t){l.e(t)}finally{l.f()}}}this._$Eh=new Map;var d,v=k(this.elementProperties);try{for(v.s();!(d=v.n()).done;){var p=f(d.value,2),y=p[0],g=p[1],m=this._$Eu(y,g);void 0!==m&&this._$Eh.set(m,y)}}catch(t){v.e(t)}finally{v.f()}this.elementStyles=this.finalizeStyles(this.styles)}}},{key:"finalizeStyles",value:function(t){var e=[];if(Array.isArray(t)){var n,r=k(new Set(t.flat(1/0).reverse()));try{for(r.s();!(n=r.n()).done;){var i=n.value;e.unshift(R(i))}}catch(t){r.e(t)}finally{r.f()}}else void 0!==t&&e.push(R(t));return e}},{key:"_$Eu",value:function(t,e){var n=e.attribute;return!1===n?void 0:"string"==typeof n?n:"string"==typeof t?t.toLowerCase():void 0}}]),e}();Y.elementStyles=[],Y.shadowRootOptions={mode:"open"},Y[F("elementProperties")]=new Map,Y[F("finalized")]=new Map,null!=G&&G({ReactiveElement:Y}),(null!==(n=B.reactiveElementVersions)&&void 0!==n?n:B.reactiveElementVersions=[]).push("2.0.4");var Z=globalThis,Q=Z.trustedTypes,X=Q?Q.createPolicy("lit-html",{createHTML:function(t){return t}}):void 0,tt="$lit$",et="lit$".concat((Math.random()+"").slice(9),"$"),nt="?"+et,rt="<".concat(nt,">"),it=document,ot=function(){return it.createComment("")},at=function(t){return null===t||"object"!=c(t)&&"function"!=typeof t},lt=Array.isArray,ut="[ \t\n\f\r]",st=/<(?:(!--|\/[^a-zA-Z])|(\/?[a-zA-Z][^>\s]*)|(\/?$))/g,ct=/-->/g,ht=/>/g,ft=RegExp(">|".concat(ut,"(?:([^\\s\"'>=/]+)(").concat(ut,"*=").concat(ut,"*(?:[^ \t\n\f\r\"'`<>=]|(\"|')|))|$)"),"g"),dt=/'/g,vt=/"/g,pt=/^(?:script|style|textarea|title)$/i,yt=Symbol.for("lit-noChange"),gt=Symbol.for("lit-nothing"),mt=new WeakMap,_t=it.createTreeWalker(it,129);function bt(t,e){if(!Array.isArray(t)||!t.hasOwnProperty("raw"))throw Error("invalid template strings array");return void 0!==X?X.createHTML(e):e}var $t=function(t,e){for(var n,r=t.length-1,i=[],o=2===e?"<svg>":"",a=st,l=0;l<r;l++){for(var u=t[l],s=void 0,c=void 0,h=-1,f=0;f<u.length&&(a.lastIndex=f,null!==(c=a.exec(u)));){var d;f=a.lastIndex,a===st?"!--"===c[1]?a=ct:void 0!==c[1]?a=ht:void 0!==c[2]?(pt.test(c[2])&&(n=RegExp("</"+c[2],"g")),a=ft):void 0!==c[3]&&(a=ft):a===ft?">"===c[0]?(a=null!==(d=n)&&void 0!==d?d:st,h=-1):void 0===c[1]?h=-2:(h=a.lastIndex-c[2].length,s=c[1],a=void 0===c[3]?ft:'"'===c[3]?vt:dt):a===vt||a===dt?a=ft:a===ct||a===ht?a=st:(a=ft,n=void 0)}var v=a===ft&&t[l+1].startsWith("/>")?" ":"";o+=a===st?u+rt:h>=0?(i.push(s),u.slice(0,h)+tt+u.slice(h)+et+v):u+et+(-2===h?l:v)}return[bt(t,o+(t[r]||"<?>")+(2===e?"</svg>":"")),i]},wt=function(){function t(e,n){var r,i=e.strings,o=e._$litType$;S(this,t),this.parts=[];var a=0,l=0,u=i.length-1,s=this.parts,c=f($t(i,o),2),d=c[0],v=c[1];if(this.el=t.createElement(d,n),_t.currentNode=this.el.content,2===o){var p=this.el.content.firstChild;p.replaceWith.apply(p,h(p.childNodes))}for(;null!==(r=_t.nextNode())&&s.length<u;){if(1===r.nodeType){if(r.hasAttributes()){var y,g=k(r.getAttributeNames());try{for(g.s();!(y=g.n()).done;){var m=y.value;if(m.endsWith(tt)){var _=v[l++],b=r.getAttribute(m).split(et),$=/([.?@])?(.*)/.exec(_);s.push({type:1,index:a,name:$[2],strings:b,ctor:"."===$[1]?Ot:"?"===$[1]?xt:"@"===$[1]?Pt:St}),r.removeAttribute(m)}else m.startsWith(et)&&(s.push({type:6,index:a}),r.removeAttribute(m))}}catch(t){g.e(t)}finally{g.f()}}if(pt.test(r.tagName)){var w=r.textContent.split(et),A=w.length-1;if(A>0){r.textContent=Q?Q.emptyScript:"";for(var E=0;E<A;E++)r.append(w[E],ot()),_t.nextNode(),s.push({type:2,index:++a});r.append(w[A],ot())}}}else if(8===r.nodeType)if(r.data===nt)s.push({type:2,index:a});else for(var O=-1;-1!==(O=r.data.indexOf(et,O+1));)s.push({type:7,index:a}),O+=et.length-1;a++}}return x(t,null,[{key:"createElement",value:function(t,e){var n=it.createElement("template");return n.innerHTML=t,n}}]),t}();function kt(t,e){var n,r,i,o,a,l=arguments.length>2&&void 0!==arguments[2]?arguments[2]:t,u=arguments.length>3?arguments[3]:void 0;if(e===yt)return e;var s=void 0!==u?null===(n=l._$Co)||void 0===n?void 0:n[u]:l._$Cl,c=at(e)?void 0:e._$litDirective$;return(null===(r=s)||void 0===r?void 0:r.constructor)!==c&&(null!==(i=s)&&void 0!==i&&null!==(o=i._$AO)&&void 0!==o&&o.call(i,!1),void 0===c?s=void 0:(s=new c(t))._$AT(t,l,u),void 0!==u?(null!==(a=l._$Co)&&void 0!==a?a:l._$Co=[])[u]=s:l._$Cl=s),void 0!==s&&(e=kt(t,s._$AS(t,e.values),s,u)),e}var At=function(){function t(e,n){S(this,t),this._$AV=[],this._$AN=void 0,this._$AD=e,this._$AM=n}return x(t,[{key:"parentNode",get:function(){return this._$AM.parentNode}},{key:"_$AU",get:function(){return this._$AM._$AU}},{key:"u",value:function(t){var e,n=this._$AD,r=n.el.content,i=n.parts,o=(null!==(e=null==t?void 0:t.creationScope)&&void 0!==e?e:it).importNode(r,!0);_t.currentNode=o;for(var a=_t.nextNode(),l=0,u=0,s=i[0];void 0!==s;){var c;if(l===s.index){var h=void 0;2===s.type?h=new Et(a,a.nextSibling,this,t):1===s.type?h=new s.ctor(a,s.name,s.strings,this,t):6===s.type&&(h=new jt(a,this,t)),this._$AV.push(h),s=i[++u]}l!==(null===(c=s)||void 0===c?void 0:c.index)&&(a=_t.nextNode(),l++)}return _t.currentNode=it,o}},{key:"p",value:function(t){var e,n=0,r=k(this._$AV);try{for(r.s();!(e=r.n()).done;){var i=e.value;void 0!==i&&(void 0!==i.strings?(i._$AI(t,i,n),n+=i.strings.length-2):i._$AI(t[n])),n++}}catch(t){r.e(t)}finally{r.f()}}}]),t}(),Et=function(){function t(e,n,r,i){var o;S(this,t),this.type=2,this._$AH=gt,this._$AN=void 0,this._$AA=e,this._$AB=n,this._$AM=r,this.options=i,this._$Cv=null===(o=null==i?void 0:i.isConnected)||void 0===o||o}return x(t,[{key:"_$AU",get:function(){var t,e;return null!==(t=null===(e=this._$AM)||void 0===e?void 0:e._$AU)&&void 0!==t?t:this._$Cv}},{key:"parentNode",get:function(){var t,e=this._$AA.parentNode,n=this._$AM;return void 0!==n&&11===(null===(t=e)||void 0===t?void 0:t.nodeType)&&(e=n.parentNode),e}},{key:"startNode",get:function(){return this._$AA}},{key:"endNode",get:function(){return this._$AB}},{key:"_$AI",value:function(t){t=kt(this,t,arguments.length>1&&void 0!==arguments[1]?arguments[1]:this),at(t)?t===gt||null==t||""===t?(this._$AH!==gt&&this._$AR(),this._$AH=gt):t!==this._$AH&&t!==yt&&this._(t):void 0!==t._$litType$?this.$(t):void 0!==t.nodeType?this.T(t):function(t){return lt(t)||"function"==typeof(null==t?void 0:t[Symbol.iterator])}(t)?this.k(t):this._(t)}},{key:"S",value:function(t){return this._$AA.parentNode.insertBefore(t,this._$AB)}},{key:"T",value:function(t){this._$AH!==t&&(this._$AR(),this._$AH=this.S(t))}},{key:"_",value:function(t){this._$AH!==gt&&at(this._$AH)?this._$AA.nextSibling.data=t:this.T(it.createTextNode(t)),this._$AH=t}},{key:"$",value:function(t){var e,n=t.values,r=t._$litType$,i="number"==typeof r?this._$AC(t):(void 0===r.el&&(r.el=wt.createElement(bt(r.h,r.h[0]),this.options)),r);if((null===(e=this._$AH)||void 0===e?void 0:e._$AD)===i)this._$AH.p(n);else{var o=new At(i,this),a=o.u(this.options);o.p(n),this.T(a),this._$AH=o}}},{key:"_$AC",value:function(t){var e=mt.get(t.strings);return void 0===e&&mt.set(t.strings,e=new wt(t)),e}},{key:"k",value:function(e){lt(this._$AH)||(this._$AH=[],this._$AR());var n,r,i=this._$AH,o=0,a=k(e);try{for(a.s();!(r=a.n()).done;){var l=r.value;o===i.length?i.push(n=new t(this.S(ot()),this.S(ot()),this,this.options)):n=i[o],n._$AI(l),o++}}catch(t){a.e(t)}finally{a.f()}o<i.length&&(this._$AR(n&&n._$AB.nextSibling,o),i.length=o)}},{key:"_$AR",value:function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:this._$AA.nextSibling,e=arguments.length>1?arguments[1]:void 0;for(null===(n=this._$AP)||void 0===n||n.call(this,!1,!0,e);t&&t!==this._$AB;){var n,r=t.nextSibling;t.remove(),t=r}}},{key:"setConnected",value:function(t){var e;void 0===this._$AM&&(this._$Cv=t,null===(e=this._$AP)||void 0===e||e.call(this,t))}}]),t}(),St=function(){function t(e,n,r,i,o){S(this,t),this.type=1,this._$AH=gt,this._$AN=void 0,this.element=e,this.name=n,this._$AM=i,this.options=o,r.length>2||""!==r[0]||""!==r[1]?(this._$AH=Array(r.length-1).fill(new String),this.strings=r):this._$AH=gt}return x(t,[{key:"tagName",get:function(){return this.element.tagName}},{key:"_$AU",get:function(){return this._$AM._$AU}},{key:"_$AI",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:this,n=arguments.length>2?arguments[2]:void 0,r=arguments.length>3?arguments[3]:void 0,i=this.strings,o=!1;if(void 0===i)t=kt(this,t,e,0),(o=!at(t)||t!==this._$AH&&t!==yt)&&(this._$AH=t);else{var a,l,u=t;for(t=i[0],a=0;a<i.length-1;a++){var s;(l=kt(this,u[n+a],e,a))===yt&&(l=this._$AH[a]),o||(o=!at(l)||l!==this._$AH[a]),l===gt?t=gt:t!==gt&&(t+=(null!==(s=l)&&void 0!==s?s:"")+i[a+1]),this._$AH[a]=l}}o&&!r&&this.j(t)}},{key:"j",value:function(t){t===gt?this.element.removeAttribute(this.name):this.element.setAttribute(this.name,null!=t?t:"")}}]),t}(),Ot=function(t){function e(){var t;return S(this,e),(t=y(this,e,arguments)).type=3,t}return m(e,St),x(e,[{key:"j",value:function(t){this.element[this.name]=t===gt?void 0:t}}]),e}(),xt=function(t){function e(){var t;return S(this,e),(t=y(this,e,arguments)).type=4,t}return m(e,St),x(e,[{key:"j",value:function(t){this.element.toggleAttribute(this.name,!!t&&t!==gt)}}]),e}(),Pt=function(t){function e(t,n,r,i,o){var a;return S(this,e),(a=y(this,e,[t,n,r,i,o])).type=5,a}return m(e,St),x(e,[{key:"_$AI",value:function(t){var e;if((t=null!==(e=kt(this,t,arguments.length>1&&void 0!==arguments[1]?arguments[1]:this,0))&&void 0!==e?e:gt)!==yt){var n=this._$AH,r=t===gt&&n!==gt||t.capture!==n.capture||t.once!==n.once||t.passive!==n.passive,i=t!==gt&&(n===gt||r);r&&this.element.removeEventListener(this.name,this,n),i&&this.element.addEventListener(this.name,this,t),this._$AH=t}}},{key:"handleEvent",value:function(t){var e,n;"function"==typeof this._$AH?this._$AH.call(null!==(e=null===(n=this.options)||void 0===n?void 0:n.host)&&void 0!==e?e:this.element,t):this._$AH.handleEvent(t)}}]),e}(),jt=function(){function t(e,n,r){S(this,t),this.element=e,this.type=6,this._$AN=void 0,this._$AM=n,this.options=r}return x(t,[{key:"_$AU",get:function(){return this._$AM._$AU}},{key:"_$AI",value:function(t){kt(this,t)}}]),t}(),Ct=Z.litHtmlPolyfillSupport;null!=Ct&&Ct(wt,Et),(null!==(r=Z.litHtmlVersions)&&void 0!==r?r:Z.litHtmlVersions=[]).push("3.1.2");var Nt=function(t){function e(){var t;return S(this,e),(t=y(this,e,arguments)).renderOptions={host:g(t)},t._$Do=void 0,t}return m(e,Y),x(e,[{key:"createRenderRoot",value:function(){var t,n,r=s(w(e.prototype),"createRenderRoot",this).call(this);return null!==(n=(t=this.renderOptions).renderBefore)&&void 0!==n||(t.renderBefore=r.firstChild),r}},{key:"update",value:function(t){var n=this.render();this.hasUpdated||(this.renderOptions.isConnected=this.isConnected),s(w(e.prototype),"update",this).call(this,t),this._$Do=function(t,e,n){var r,i=null!==(r=null==n?void 0:n.renderBefore)&&void 0!==r?r:e,o=i._$litPart$;if(void 0===o){var a,l=null!==(a=null==n?void 0:n.renderBefore)&&void 0!==a?a:null;i._$litPart$=o=new Et(e.insertBefore(ot(),l),l,void 0,null!=n?n:{})}return o._$AI(t),o}(n,this.renderRoot,this.renderOptions)}},{key:"connectedCallback",value:function(){var t;s(w(e.prototype),"connectedCallback",this).call(this),null===(t=this._$Do)||void 0===t||t.setConnected(!0)}},{key:"disconnectedCallback",value:function(){var t;s(w(e.prototype),"disconnectedCallback",this).call(this),null===(t=this._$Do)||void 0===t||t.setConnected(!1)}},{key:"render",value:function(){return yt}}]),e}();Nt._$litElement$=!0,Nt.finalized=!0,null===(i=globalThis.litElementHydrateSupport)||void 0===i||i.call(globalThis,{LitElement:Nt});var Lt=globalThis.litElementPolyfillSupport;null==Lt||Lt({LitElement:Nt}),(null!==(o=globalThis.litElementVersions)&&void 0!==o?o:globalThis.litElementVersions=[]).push("4.0.4");var Tt="code",Rt="pin_used",Ut="pin_synced_to_locks",It=["number_of_uses"],Mt=["name","enabled","pin",Ut].concat(It,[Tt,Rt]),Ht="fold-entity-row.js",Dt={type:"divider"};function zt(t,e,n,r,i){return Bt.apply(this,arguments)}function Bt(){return Bt=p(d().mark((function t(e,n,r,i,o){var a,l,u,s,c,v,p,y,g,m;return d().wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,Promise.all([e.callWS({config_entry_id:n,type:"lock_code_manager/get_slot_calendar_data"}),e.callWS({type:"lovelace/resources"})]);case 2:return a=t.sent,l=f(a,2),u=l[0],s=l[1],c=i.map((function(t){return Vt(t)})).sort(Wt),v=Object.keys(u.slots).map((function(t){return parseInt(t,10)})),p=v.map((function(t){return qt(e,t,c,u)})),y=[].concat(h(u.locks.sort((function(t,e){return t.localeCompare(e)}))),h(c.filter((function(t){return"pin_synced_to_locks"===t.key})).map((function(t){return{entity:t.entity_id,name:(t.name?t.name:t.original_name).replace("PIN synced to locks","synced").replace("Code slot","Slot"),type:"state-label"}})))),g=s.filter((function(t){return t.url.includes(Ht)})).length>0,m=p.map((function(t){return Ft(t,g,o)})),t.abrupt("return",{badges:y,cards:m,panel:!1,path:Kt(r),title:r});case 13:case"end":return t.stop()}}),t)}))),Bt.apply(this,arguments)}function Wt(t,e){return t.slotNum<e.slotNum?-1:t.slotNum>e.slotNum?1:Mt.indexOf(t.key)<Mt.indexOf(e.key)?-1:Mt.indexOf(t.key)>Mt.indexOf(e.key)?1:t.key===e.key&&[Rt,Tt].includes(t.key)&&t.lockEntityId<e.lockEntityId?-1:1}function Vt(t){var e=t.unique_id.split("|");return l(l({},t),{},{key:e[2],lockEntityId:e[3],slotNum:parseInt(e[1],10)})}function Gt(t){return t.map((function(t){return{entity:t}}))}function Ft(t,e,n){return{cards:[{content:"## Code Slot ".concat(t.slotNum),type:"markdown"},{entities:[].concat(h(Gt(t.mainEntityIds)),[Dt,{entity:t.pinShouldBeEnabledEntity.entity_id,name:"PIN synced to locks"},{entity:t.codeEventEntityId,name:"PIN Last Used"}],h(Jt(t.conditionEntityIds,"Conditions",e)),h(n?Jt(t.codeSensorEntityIds,"Code Slot Sensors",e):[])),show_header_toggle:!1,type:"entities"}],type:"vertical-stack"}}function qt(t,e,n,r){var i,o=[],a=[],l=[];n.filter((function(t){return t.slotNum===e})).forEach((function(t){t.key===Tt?l.push(t.entity_id):t.key===Rt?i=t.entity_id:It.includes(t.key)?a.push(t.entity_id):t.key!==Ut&&o.push(t.entity_id)}));var u=n.find((function(t){return t.slotNum===e&&t.key===Ut})),s=r.slots[e];return s&&a.unshift(s),{codeEventEntityId:i,codeSensorEntityIds:l,conditionEntityIds:a,mainEntityIds:o,pinShouldBeEnabledEntity:u,slotNum:e}}function Jt(t,e,n){if(0===t.length)return[];var r=Gt(t);return n?[Dt,{entities:r,head:{label:e,type:"section"},type:"custom:fold-entity-row"}]:[{label:e,type:"section"}].concat(h(r))}function Kt(t){var e,n=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"-",r="àáâäæãåāăąçćčđďèéêëēėęěğǵḧîïíīįìıİłḿñńǹňôöòóœøōõőṕŕřßśšşșťțûüùúūǘůűųẃẍÿýžźż·",i="aaaaaaaaaacccddeeeeeeeegghiiiiiiiilmnnnnoooooooooprrsssssttuuuuuuuuuwxyyzzz".concat(n),o=new RegExp(r.split("").join("|"),"g");return""===t?e="":(e=t.toString().toLowerCase().replace(o,(function(t){return i.charAt(r.indexOf(t))})).replace(/(\d),(?=\d)/g,"$1").replace(/[^a-z0-9]+/g,n).replace(new RegExp("(".concat(n,")\\1+"),"g"),"$1").replace(new RegExp("^".concat(n,"+")),"").replace(new RegExp("".concat(n,"+$")),""),""===e&&(e="unknown")),e}var Yt=function(t){function e(){return S(this,e),y(this,e,arguments)}var n;return m(e,Y),x(e,null,[{key:"generate",value:(n=p(d().mark((function t(e,n){var r,i;return d().wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,n.callWS({domain:"lock_code_manager",type:"config_entries/get"});case 2:if(0!==(r=t.sent).length){t.next=5;break}return t.abrupt("return",{title:"Lock Code Manager",views:[{badges:[],cards:[{content:"# No Lock Code Manager configurations found!",type:"markdown"}],title:"Lock Code Manager"}]});case 5:return t.next=7,Promise.all(r.map((function(t){var n;return{path:Kt(t.title),strategy:{config_entry_id:t.entry_id,include_code_slot_sensors:null!==(n=e.include_code_slot_sensors)&&void 0!==n&&n,type:"custom:lock-code-manager"},title:t.title}})));case 7:return 1===(i=t.sent).length&&i.push({title:"​"}),t.abrupt("return",{title:"Lock Code Manager",views:i});case 10:case"end":return t.stop()}}),t)}))),function(t,e){return n.apply(this,arguments)})}]),e}(),Zt=function(t){function e(){return S(this,e),y(this,e,arguments)}var n;return m(e,Y),x(e,null,[{key:"generate",value:(n=p(d().mark((function t(e,n){var r,i,o,a,l,u,s,c,h;return d().wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(r=e.config_entry_id,i=e.config_entry_title,!(void 0===r&&void 0===i||void 0!==r&&void 0!==i)){t.next=3;break}return t.abrupt("return",{badges:[],cards:[{content:"## ERROR: Either `config_entry_title` or `config_entry_id` must be provided in the view config, but not both!",type:"markdown"}],title:"Lock Code Manager"});case 3:return t.prev=3,t.next=6,n.callWS({config_entry_id:r,config_entry_title:i,type:"lock_code_manager/get_config_entry_entities"});case 6:return a=t.sent,l=f(a,3),u=l[0],s=l[1],c=l[2],t.abrupt("return",zt(n,u,s,c,null!==(o=e.include_code_slot_sensors)&&void 0!==o&&o));case 14:return t.prev=14,t.t0=t.catch(3),h=void 0!==r?"with ID `".concat(r,"`"):"called `".concat(i,"`"),t.abrupt("return",{badges:[],cards:[{content:"## ERROR: No Lock Code Manager configuration ".concat(h," found!"),type:"markdown"}],title:"Lock Code Manager"});case 18:case"end":return t.stop()}}),t,null,[[3,14]])}))),function(t,e){return n.apply(this,arguments)})}]),e}();customElements.define("ll-strategy-dashboard-lock-code-manager",Yt),customElements.define("ll-strategy-view-lock-code-manager",Zt);
+var t,e,n,r,i,o;function a(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function l(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?a(Object(n),!0).forEach((function(e){u(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):a(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function u(t,e,n){return(e=P(e))in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}function s(){return s="undefined"!=typeof Reflect&&Reflect.get?Reflect.get.bind():function(t,e,n){var r=function(t,e){for(;!Object.prototype.hasOwnProperty.call(t,e)&&null!==(t=w(t)););return t}(t,e);if(r){var i=Object.getOwnPropertyDescriptor(r,e);return i.get?i.get.call(arguments.length<3?t:n):i.value}},s.apply(this,arguments)}function c(t){return c="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t},c(t)}function h(t){return function(t){if(Array.isArray(t))return E(t)}(t)||function(t){if("undefined"!=typeof Symbol&&null!=t[Symbol.iterator]||null!=t["@@iterator"])return Array.from(t)}(t)||A(t)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function f(t,e){return function(t){if(Array.isArray(t))return t}(t)||function(t,e){var n=null==t?null:"undefined"!=typeof Symbol&&t[Symbol.iterator]||t["@@iterator"];if(null!=n){var r,i,o,a,l=[],u=!0,s=!1;try{if(o=(n=n.call(t)).next,0===e){if(Object(n)!==n)return;u=!1}else for(;!(u=(r=o.call(n)).done)&&(l.push(r.value),l.length!==e);u=!0);}catch(t){s=!0,i=t}finally{try{if(!u&&null!=n.return&&(a=n.return(),Object(a)!==a))return}finally{if(s)throw i}}return l}}(t,e)||A(t,e)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function d(){d=function(){return e};var t,e={},n=Object.prototype,r=n.hasOwnProperty,i=Object.defineProperty||function(t,e,n){t[e]=n.value},o="function"==typeof Symbol?Symbol:{},a=o.iterator||"@@iterator",l=o.asyncIterator||"@@asyncIterator",u=o.toStringTag||"@@toStringTag";function s(t,e,n){return Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}),t[e]}try{s({},"")}catch(t){s=function(t,e,n){return t[e]=n}}function h(t,e,n,r){var o=e&&e.prototype instanceof _?e:_,a=Object.create(o.prototype),l=new N(r||[]);return i(a,"_invoke",{value:x(t,n,l)}),a}function f(t,e,n){try{return{type:"normal",arg:t.call(e,n)}}catch(t){return{type:"throw",arg:t}}}e.wrap=h;var v="suspendedStart",p="suspendedYield",y="executing",g="completed",m={};function _(){}function b(){}function $(){}var w={};s(w,a,(function(){return this}));var k=Object.getPrototypeOf,A=k&&k(k(L([])));A&&A!==n&&r.call(A,a)&&(w=A);var E=$.prototype=_.prototype=Object.create(w);function S(t){["next","throw","return"].forEach((function(e){s(t,e,(function(t){return this._invoke(e,t)}))}))}function O(t,e){function n(i,o,a,l){var u=f(t[i],t,o);if("throw"!==u.type){var s=u.arg,h=s.value;return h&&"object"==c(h)&&r.call(h,"__await")?e.resolve(h.__await).then((function(t){n("next",t,a,l)}),(function(t){n("throw",t,a,l)})):e.resolve(h).then((function(t){s.value=t,a(s)}),(function(t){return n("throw",t,a,l)}))}l(u.arg)}var o;i(this,"_invoke",{value:function(t,r){function i(){return new e((function(e,i){n(t,r,e,i)}))}return o=o?o.then(i,i):i()}})}function x(e,n,r){var i=v;return function(o,a){if(i===y)throw new Error("Generator is already running");if(i===g){if("throw"===o)throw a;return{value:t,done:!0}}for(r.method=o,r.arg=a;;){var l=r.delegate;if(l){var u=P(l,r);if(u){if(u===m)continue;return u}}if("next"===r.method)r.sent=r._sent=r.arg;else if("throw"===r.method){if(i===v)throw i=g,r.arg;r.dispatchException(r.arg)}else"return"===r.method&&r.abrupt("return",r.arg);i=y;var s=f(e,n,r);if("normal"===s.type){if(i=r.done?g:p,s.arg===m)continue;return{value:s.arg,done:r.done}}"throw"===s.type&&(i=g,r.method="throw",r.arg=s.arg)}}}function P(e,n){var r=n.method,i=e.iterator[r];if(i===t)return n.delegate=null,"throw"===r&&e.iterator.return&&(n.method="return",n.arg=t,P(e,n),"throw"===n.method)||"return"!==r&&(n.method="throw",n.arg=new TypeError("The iterator does not provide a '"+r+"' method")),m;var o=f(i,e.iterator,n.arg);if("throw"===o.type)return n.method="throw",n.arg=o.arg,n.delegate=null,m;var a=o.arg;return a?a.done?(n[e.resultName]=a.value,n.next=e.nextLoc,"return"!==n.method&&(n.method="next",n.arg=t),n.delegate=null,m):a:(n.method="throw",n.arg=new TypeError("iterator result is not an object"),n.delegate=null,m)}function j(t){var e={tryLoc:t[0]};1 in t&&(e.catchLoc=t[1]),2 in t&&(e.finallyLoc=t[2],e.afterLoc=t[3]),this.tryEntries.push(e)}function C(t){var e=t.completion||{};e.type="normal",delete e.arg,t.completion=e}function N(t){this.tryEntries=[{tryLoc:"root"}],t.forEach(j,this),this.reset(!0)}function L(e){if(e||""===e){var n=e[a];if(n)return n.call(e);if("function"==typeof e.next)return e;if(!isNaN(e.length)){var i=-1,o=function n(){for(;++i<e.length;)if(r.call(e,i))return n.value=e[i],n.done=!1,n;return n.value=t,n.done=!0,n};return o.next=o}}throw new TypeError(c(e)+" is not iterable")}return b.prototype=$,i(E,"constructor",{value:$,configurable:!0}),i($,"constructor",{value:b,configurable:!0}),b.displayName=s($,u,"GeneratorFunction"),e.isGeneratorFunction=function(t){var e="function"==typeof t&&t.constructor;return!!e&&(e===b||"GeneratorFunction"===(e.displayName||e.name))},e.mark=function(t){return Object.setPrototypeOf?Object.setPrototypeOf(t,$):(t.__proto__=$,s(t,u,"GeneratorFunction")),t.prototype=Object.create(E),t},e.awrap=function(t){return{__await:t}},S(O.prototype),s(O.prototype,l,(function(){return this})),e.AsyncIterator=O,e.async=function(t,n,r,i,o){void 0===o&&(o=Promise);var a=new O(h(t,n,r,i),o);return e.isGeneratorFunction(n)?a:a.next().then((function(t){return t.done?t.value:a.next()}))},S(E),s(E,u,"Generator"),s(E,a,(function(){return this})),s(E,"toString",(function(){return"[object Generator]"})),e.keys=function(t){var e=Object(t),n=[];for(var r in e)n.push(r);return n.reverse(),function t(){for(;n.length;){var r=n.pop();if(r in e)return t.value=r,t.done=!1,t}return t.done=!0,t}},e.values=L,N.prototype={constructor:N,reset:function(e){if(this.prev=0,this.next=0,this.sent=this._sent=t,this.done=!1,this.delegate=null,this.method="next",this.arg=t,this.tryEntries.forEach(C),!e)for(var n in this)"t"===n.charAt(0)&&r.call(this,n)&&!isNaN(+n.slice(1))&&(this[n]=t)},stop:function(){this.done=!0;var t=this.tryEntries[0].completion;if("throw"===t.type)throw t.arg;return this.rval},dispatchException:function(e){if(this.done)throw e;var n=this;function i(r,i){return l.type="throw",l.arg=e,n.next=r,i&&(n.method="next",n.arg=t),!!i}for(var o=this.tryEntries.length-1;o>=0;--o){var a=this.tryEntries[o],l=a.completion;if("root"===a.tryLoc)return i("end");if(a.tryLoc<=this.prev){var u=r.call(a,"catchLoc"),s=r.call(a,"finallyLoc");if(u&&s){if(this.prev<a.catchLoc)return i(a.catchLoc,!0);if(this.prev<a.finallyLoc)return i(a.finallyLoc)}else if(u){if(this.prev<a.catchLoc)return i(a.catchLoc,!0)}else{if(!s)throw new Error("try statement without catch or finally");if(this.prev<a.finallyLoc)return i(a.finallyLoc)}}}},abrupt:function(t,e){for(var n=this.tryEntries.length-1;n>=0;--n){var i=this.tryEntries[n];if(i.tryLoc<=this.prev&&r.call(i,"finallyLoc")&&this.prev<i.finallyLoc){var o=i;break}}o&&("break"===t||"continue"===t)&&o.tryLoc<=e&&e<=o.finallyLoc&&(o=null);var a=o?o.completion:{};return a.type=t,a.arg=e,o?(this.method="next",this.next=o.finallyLoc,m):this.complete(a)},complete:function(t,e){if("throw"===t.type)throw t.arg;return"break"===t.type||"continue"===t.type?this.next=t.arg:"return"===t.type?(this.rval=this.arg=t.arg,this.method="return",this.next="end"):"normal"===t.type&&e&&(this.next=e),m},finish:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var n=this.tryEntries[e];if(n.finallyLoc===t)return this.complete(n.completion,n.afterLoc),C(n),m}},catch:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var n=this.tryEntries[e];if(n.tryLoc===t){var r=n.completion;if("throw"===r.type){var i=r.arg;C(n)}return i}}throw new Error("illegal catch attempt")},delegateYield:function(e,n,r){return this.delegate={iterator:L(e),resultName:n,nextLoc:r},"next"===this.method&&(this.arg=t),m}},e}function v(t,e,n,r,i,o,a){try{var l=t[o](a),u=l.value}catch(t){return void n(t)}l.done?e(u):Promise.resolve(u).then(r,i)}function p(t){return function(){var e=this,n=arguments;return new Promise((function(r,i){var o=t.apply(e,n);function a(t){v(o,r,i,a,l,"next",t)}function l(t){v(o,r,i,a,l,"throw",t)}a(void 0)}))}}function y(t,e,n){return e=w(e),function(t,e){if(e&&("object"===c(e)||"function"==typeof e))return e;if(void 0!==e)throw new TypeError("Derived constructors may only return object or undefined");return g(t)}(t,b()?Reflect.construct(e,n||[],w(t).constructor):e.apply(t,n))}function g(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}function m(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Super expression must either be null or a function");t.prototype=Object.create(e&&e.prototype,{constructor:{value:t,writable:!0,configurable:!0}}),Object.defineProperty(t,"prototype",{writable:!1}),e&&$(t,e)}function _(t){var e="function"==typeof Map?new Map:void 0;return _=function(t){if(null===t||!function(t){try{return-1!==Function.toString.call(t).indexOf("[native code]")}catch(e){return"function"==typeof t}}(t))return t;if("function"!=typeof t)throw new TypeError("Super expression must either be null or a function");if(void 0!==e){if(e.has(t))return e.get(t);e.set(t,n)}function n(){return function(t,e,n){if(b())return Reflect.construct.apply(null,arguments);var r=[null];r.push.apply(r,e);var i=new(t.bind.apply(t,r));return n&&$(i,n.prototype),i}(t,arguments,w(this).constructor)}return n.prototype=Object.create(t.prototype,{constructor:{value:n,enumerable:!1,writable:!0,configurable:!0}}),$(n,t)},_(t)}function b(){try{var t=!Boolean.prototype.valueOf.call(Reflect.construct(Boolean,[],(function(){})))}catch(t){}return(b=function(){return!!t})()}function $(t,e){return $=Object.setPrototypeOf?Object.setPrototypeOf.bind():function(t,e){return t.__proto__=e,t},$(t,e)}function w(t){return w=Object.setPrototypeOf?Object.getPrototypeOf.bind():function(t){return t.__proto__||Object.getPrototypeOf(t)},w(t)}function k(t,e){var n="undefined"!=typeof Symbol&&t[Symbol.iterator]||t["@@iterator"];if(!n){if(Array.isArray(t)||(n=A(t))||e&&t&&"number"==typeof t.length){n&&(t=n);var r=0,i=function(){};return{s:i,n:function(){return r>=t.length?{done:!0}:{done:!1,value:t[r++]}},e:function(t){throw t},f:i}}throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}var o,a=!0,l=!1;return{s:function(){n=n.call(t)},n:function(){var t=n.next();return a=t.done,t},e:function(t){l=!0,o=t},f:function(){try{a||null==n.return||n.return()}finally{if(l)throw o}}}}function A(t,e){if(t){if("string"==typeof t)return E(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);return"Object"===n&&t.constructor&&(n=t.constructor.name),"Map"===n||"Set"===n?Array.from(t):"Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)?E(t,e):void 0}}function E(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}function S(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}function O(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,P(r.key),r)}}function x(t,e,n){return e&&O(t.prototype,e),n&&O(t,n),Object.defineProperty(t,"prototype",{writable:!1}),t}function P(t){var e=function(t,e){if("object"!=c(t)||!t)return t;var n=t[Symbol.toPrimitive];if(void 0!==n){var r=n.call(t,e||"default");if("object"!=c(r))return r;throw new TypeError("@@toPrimitive must return a primitive value.")}return("string"===e?String:Number)(t)}(t,"string");return"symbol"==c(e)?e:String(e)}var j=globalThis,C=j.ShadowRoot&&(void 0===j.ShadyCSS||j.ShadyCSS.nativeShadow)&&"adoptedStyleSheets"in Document.prototype&&"replace"in CSSStyleSheet.prototype,N=Symbol(),L=new WeakMap,T=function(){function t(e,n,r){if(S(this,t),this._$cssResult$=!0,r!==N)throw Error("CSSResult is not constructable. Use `unsafeCSS` or `css` instead.");this.cssText=e,this.t=n}return x(t,[{key:"styleSheet",get:function(){var t=this.o,e=this.t;if(C&&void 0===t){var n=void 0!==e&&1===e.length;n&&(t=L.get(e)),void 0===t&&((this.o=t=new CSSStyleSheet).replaceSync(this.cssText),n&&L.set(e,t))}return t}},{key:"toString",value:function(){return this.cssText}}]),t}(),R=C?function(t){return t}:function(t){return t instanceof CSSStyleSheet?function(t){var e,n="",r=k(t.cssRules);try{for(r.s();!(e=r.n()).done;){n+=e.value.cssText}}catch(t){r.e(t)}finally{r.f()}return function(t){return new T("string"==typeof t?t:t+"",void 0,N)}(n)}(t):t},U=Object.is,I=Object.defineProperty,M=Object.getOwnPropertyDescriptor,H=Object.getOwnPropertyNames,D=Object.getOwnPropertySymbols,z=Object.getPrototypeOf,B=globalThis,W=B.trustedTypes,V=W?W.emptyScript:"",G=B.reactiveElementPolyfillSupport,F=function(t,e){return t},q={toAttribute:function(t,e){switch(e){case Boolean:t=t?V:null;break;case Object:case Array:t=null==t?t:JSON.stringify(t)}return t},fromAttribute:function(t,e){var n=t;switch(e){case Boolean:n=null!==t;break;case Number:n=null===t?null:Number(t);break;case Object:case Array:try{n=JSON.parse(t)}catch(t){n=null}}return n}},J=function(t,e){return!U(t,e)},K={attribute:!0,type:String,converter:q,reflect:!1,hasChanged:J};null!==(t=Symbol.metadata)&&void 0!==t||(Symbol.metadata=Symbol("metadata")),null!==(e=B.litPropertyMetadata)&&void 0!==e||(B.litPropertyMetadata=new WeakMap);var Y=function(t){function e(){var t;return S(this,e),(t=y(this,e))._$Ep=void 0,t.isUpdatePending=!1,t.hasUpdated=!1,t._$Em=null,t._$Ev(),t}var n;return m(e,_(HTMLElement)),x(e,[{key:"_$Ev",value:function(){var t,e=this;this._$ES=new Promise((function(t){return e.enableUpdating=t})),this._$AL=new Map,this._$E_(),this.requestUpdate(),null===(t=this.constructor.l)||void 0===t||t.forEach((function(t){return t(e)}))}},{key:"addController",value:function(t){var e,n;(null!==(e=this._$EO)&&void 0!==e?e:this._$EO=new Set).add(t),void 0!==this.renderRoot&&this.isConnected&&(null===(n=t.hostConnected)||void 0===n||n.call(t))}},{key:"removeController",value:function(t){var e;null===(e=this._$EO)||void 0===e||e.delete(t)}},{key:"_$E_",value:function(){var t,e=new Map,n=k(this.constructor.elementProperties.keys());try{for(n.s();!(t=n.n()).done;){var r=t.value;this.hasOwnProperty(r)&&(e.set(r,this[r]),delete this[r])}}catch(t){n.e(t)}finally{n.f()}e.size>0&&(this._$Ep=e)}},{key:"createRenderRoot",value:function(){var t,e=null!==(t=this.shadowRoot)&&void 0!==t?t:this.attachShadow(this.constructor.shadowRootOptions);return function(t,e){if(C)t.adoptedStyleSheets=e.map((function(t){return t instanceof CSSStyleSheet?t:t.styleSheet}));else{var n,r=k(e);try{for(r.s();!(n=r.n()).done;){var i=n.value,o=document.createElement("style"),a=j.litNonce;void 0!==a&&o.setAttribute("nonce",a),o.textContent=i.cssText,t.appendChild(o)}}catch(t){r.e(t)}finally{r.f()}}}(e,this.constructor.elementStyles),e}},{key:"connectedCallback",value:function(){var t,e;null!==(t=this.renderRoot)&&void 0!==t||(this.renderRoot=this.createRenderRoot()),this.enableUpdating(!0),null===(e=this._$EO)||void 0===e||e.forEach((function(t){var e;return null===(e=t.hostConnected)||void 0===e?void 0:e.call(t)}))}},{key:"enableUpdating",value:function(t){}},{key:"disconnectedCallback",value:function(){var t;null===(t=this._$EO)||void 0===t||t.forEach((function(t){var e;return null===(e=t.hostDisconnected)||void 0===e?void 0:e.call(t)}))}},{key:"attributeChangedCallback",value:function(t,e,n){this._$AK(t,n)}},{key:"_$EC",value:function(t,e){var n=this.constructor.elementProperties.get(t),r=this.constructor._$Eu(t,n);if(void 0!==r&&!0===n.reflect){var i,o=(void 0!==(null===(i=n.converter)||void 0===i?void 0:i.toAttribute)?n.converter:q).toAttribute(e,n.type);this._$Em=t,null==o?this.removeAttribute(r):this.setAttribute(r,o),this._$Em=null}}},{key:"_$AK",value:function(t,e){var n=this.constructor,r=n._$Eh.get(t);if(void 0!==r&&this._$Em!==r){var i,o=n.getPropertyOptions(r),a="function"==typeof o.converter?{fromAttribute:o.converter}:void 0!==(null===(i=o.converter)||void 0===i?void 0:i.fromAttribute)?o.converter:q;this._$Em=r,this[r]=a.fromAttribute(e,o.type),this._$Em=null}}},{key:"requestUpdate",value:function(t,e,n){if(void 0!==t){var r,i;if(null!==(r=n)&&void 0!==r||(n=this.constructor.getPropertyOptions(t)),!(null!==(i=n.hasChanged)&&void 0!==i?i:J)(this[t],e))return;this.P(t,e,n)}!1===this.isUpdatePending&&(this._$ES=this._$ET())}},{key:"P",value:function(t,e,n){var r;this._$AL.has(t)||this._$AL.set(t,e),!0===n.reflect&&this._$Em!==t&&(null!==(r=this._$Ej)&&void 0!==r?r:this._$Ej=new Set).add(t)}},{key:"_$ET",value:(n=p(d().mark((function t(){var e;return d().wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return this.isUpdatePending=!0,t.prev=1,t.next=4,this._$ES;case 4:t.next=9;break;case 6:t.prev=6,t.t0=t.catch(1),Promise.reject(t.t0);case 9:if(e=this.scheduleUpdate(),t.t1=null!=e,!t.t1){t.next=14;break}return t.next=14,e;case 14:return t.abrupt("return",!this.isUpdatePending);case 15:case"end":return t.stop()}}),t,this,[[1,6]])}))),function(){return n.apply(this,arguments)})},{key:"scheduleUpdate",value:function(){return this.performUpdate()}},{key:"performUpdate",value:function(){if(this.isUpdatePending){if(!this.hasUpdated){var t;if(null!==(t=this.renderRoot)&&void 0!==t||(this.renderRoot=this.createRenderRoot()),this._$Ep){var e,n=k(this._$Ep);try{for(n.s();!(e=n.n()).done;){var r=f(e.value,2),i=r[0],o=r[1];this[i]=o}}catch(t){n.e(t)}finally{n.f()}this._$Ep=void 0}var a=this.constructor.elementProperties;if(a.size>0){var l,u=k(a);try{for(u.s();!(l=u.n()).done;){var s=f(l.value,2),c=s[0],h=s[1];!0!==h.wrapped||this._$AL.has(c)||void 0===this[c]||this.P(c,this[c],h)}}catch(t){u.e(t)}finally{u.f()}}}var d=!1,v=this._$AL;try{var p;(d=this.shouldUpdate(v))?(this.willUpdate(v),null!==(p=this._$EO)&&void 0!==p&&p.forEach((function(t){var e;return null===(e=t.hostUpdate)||void 0===e?void 0:e.call(t)})),this.update(v)):this._$EU()}catch(v){throw d=!1,this._$EU(),v}d&&this._$AE(v)}}},{key:"willUpdate",value:function(t){}},{key:"_$AE",value:function(t){var e;null!==(e=this._$EO)&&void 0!==e&&e.forEach((function(t){var e;return null===(e=t.hostUpdated)||void 0===e?void 0:e.call(t)})),this.hasUpdated||(this.hasUpdated=!0,this.firstUpdated(t)),this.updated(t)}},{key:"_$EU",value:function(){this._$AL=new Map,this.isUpdatePending=!1}},{key:"updateComplete",get:function(){return this.getUpdateComplete()}},{key:"getUpdateComplete",value:function(){return this._$ES}},{key:"shouldUpdate",value:function(t){return!0}},{key:"update",value:function(t){var e=this;this._$Ej&&(this._$Ej=this._$Ej.forEach((function(t){return e._$EC(t,e[t])}))),this._$EU()}},{key:"updated",value:function(t){}},{key:"firstUpdated",value:function(t){}}],[{key:"addInitializer",value:function(t){var e;this._$Ei(),(null!==(e=this.l)&&void 0!==e?e:this.l=[]).push(t)}},{key:"observedAttributes",get:function(){return this.finalize(),this._$Eh&&h(this._$Eh.keys())}},{key:"createProperty",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:K;if(e.state&&(e.attribute=!1),this._$Ei(),this.elementProperties.set(t,e),!e.noAccessor){var n=Symbol(),r=this.getPropertyDescriptor(t,n,e);void 0!==r&&I(this.prototype,t,r)}}},{key:"getPropertyDescriptor",value:function(t,e,n){var r,i=null!==(r=M(this.prototype,t))&&void 0!==r?r:{get:function(){return this[e]},set:function(t){this[e]=t}},o=i.get,a=i.set;return{get:function(){return null==o?void 0:o.call(this)},set:function(e){var r=null==o?void 0:o.call(this);a.call(this,e),this.requestUpdate(t,r,n)},configurable:!0,enumerable:!0}}},{key:"getPropertyOptions",value:function(t){var e;return null!==(e=this.elementProperties.get(t))&&void 0!==e?e:K}},{key:"_$Ei",value:function(){if(!this.hasOwnProperty(F("elementProperties"))){var t=z(this);t.finalize(),void 0!==t.l&&(this.l=h(t.l)),this.elementProperties=new Map(t.elementProperties)}}},{key:"finalize",value:function(){if(!this.hasOwnProperty(F("finalized"))){if(this.finalized=!0,this._$Ei(),this.hasOwnProperty(F("properties"))){var t,e=this.properties,n=k([].concat(h(H(e)),h(D(e))));try{for(n.s();!(t=n.n()).done;){var r=t.value;this.createProperty(r,e[r])}}catch(t){n.e(t)}finally{n.f()}}var i=this[Symbol.metadata];if(null!==i){var o=litPropertyMetadata.get(i);if(void 0!==o){var a,l=k(o);try{for(l.s();!(a=l.n()).done;){var u=f(a.value,2),s=u[0],c=u[1];this.elementProperties.set(s,c)}}catch(t){l.e(t)}finally{l.f()}}}this._$Eh=new Map;var d,v=k(this.elementProperties);try{for(v.s();!(d=v.n()).done;){var p=f(d.value,2),y=p[0],g=p[1],m=this._$Eu(y,g);void 0!==m&&this._$Eh.set(m,y)}}catch(t){v.e(t)}finally{v.f()}this.elementStyles=this.finalizeStyles(this.styles)}}},{key:"finalizeStyles",value:function(t){var e=[];if(Array.isArray(t)){var n,r=k(new Set(t.flat(1/0).reverse()));try{for(r.s();!(n=r.n()).done;){var i=n.value;e.unshift(R(i))}}catch(t){r.e(t)}finally{r.f()}}else void 0!==t&&e.push(R(t));return e}},{key:"_$Eu",value:function(t,e){var n=e.attribute;return!1===n?void 0:"string"==typeof n?n:"string"==typeof t?t.toLowerCase():void 0}}]),e}();Y.elementStyles=[],Y.shadowRootOptions={mode:"open"},Y[F("elementProperties")]=new Map,Y[F("finalized")]=new Map,null!=G&&G({ReactiveElement:Y}),(null!==(n=B.reactiveElementVersions)&&void 0!==n?n:B.reactiveElementVersions=[]).push("2.0.4");var Z=globalThis,Q=Z.trustedTypes,X=Q?Q.createPolicy("lit-html",{createHTML:function(t){return t}}):void 0,tt="$lit$",et="lit$".concat((Math.random()+"").slice(9),"$"),nt="?"+et,rt="<".concat(nt,">"),it=document,ot=function(){return it.createComment("")},at=function(t){return null===t||"object"!=c(t)&&"function"!=typeof t},lt=Array.isArray,ut="[ \t\n\f\r]",st=/<(?:(!--|\/[^a-zA-Z])|(\/?[a-zA-Z][^>\s]*)|(\/?$))/g,ct=/-->/g,ht=/>/g,ft=RegExp(">|".concat(ut,"(?:([^\\s\"'>=/]+)(").concat(ut,"*=").concat(ut,"*(?:[^ \t\n\f\r\"'`<>=]|(\"|')|))|$)"),"g"),dt=/'/g,vt=/"/g,pt=/^(?:script|style|textarea|title)$/i,yt=Symbol.for("lit-noChange"),gt=Symbol.for("lit-nothing"),mt=new WeakMap,_t=it.createTreeWalker(it,129);function bt(t,e){if(!Array.isArray(t)||!t.hasOwnProperty("raw"))throw Error("invalid template strings array");return void 0!==X?X.createHTML(e):e}var $t=function(t,e){for(var n,r=t.length-1,i=[],o=2===e?"<svg>":"",a=st,l=0;l<r;l++){for(var u=t[l],s=void 0,c=void 0,h=-1,f=0;f<u.length&&(a.lastIndex=f,null!==(c=a.exec(u)));){var d;f=a.lastIndex,a===st?"!--"===c[1]?a=ct:void 0!==c[1]?a=ht:void 0!==c[2]?(pt.test(c[2])&&(n=RegExp("</"+c[2],"g")),a=ft):void 0!==c[3]&&(a=ft):a===ft?">"===c[0]?(a=null!==(d=n)&&void 0!==d?d:st,h=-1):void 0===c[1]?h=-2:(h=a.lastIndex-c[2].length,s=c[1],a=void 0===c[3]?ft:'"'===c[3]?vt:dt):a===vt||a===dt?a=ft:a===ct||a===ht?a=st:(a=ft,n=void 0)}var v=a===ft&&t[l+1].startsWith("/>")?" ":"";o+=a===st?u+rt:h>=0?(i.push(s),u.slice(0,h)+tt+u.slice(h)+et+v):u+et+(-2===h?l:v)}return[bt(t,o+(t[r]||"<?>")+(2===e?"</svg>":"")),i]},wt=function(){function t(e,n){var r,i=e.strings,o=e._$litType$;S(this,t),this.parts=[];var a=0,l=0,u=i.length-1,s=this.parts,c=f($t(i,o),2),d=c[0],v=c[1];if(this.el=t.createElement(d,n),_t.currentNode=this.el.content,2===o){var p=this.el.content.firstChild;p.replaceWith.apply(p,h(p.childNodes))}for(;null!==(r=_t.nextNode())&&s.length<u;){if(1===r.nodeType){if(r.hasAttributes()){var y,g=k(r.getAttributeNames());try{for(g.s();!(y=g.n()).done;){var m=y.value;if(m.endsWith(tt)){var _=v[l++],b=r.getAttribute(m).split(et),$=/([.?@])?(.*)/.exec(_);s.push({type:1,index:a,name:$[2],strings:b,ctor:"."===$[1]?Ot:"?"===$[1]?xt:"@"===$[1]?Pt:St}),r.removeAttribute(m)}else m.startsWith(et)&&(s.push({type:6,index:a}),r.removeAttribute(m))}}catch(t){g.e(t)}finally{g.f()}}if(pt.test(r.tagName)){var w=r.textContent.split(et),A=w.length-1;if(A>0){r.textContent=Q?Q.emptyScript:"";for(var E=0;E<A;E++)r.append(w[E],ot()),_t.nextNode(),s.push({type:2,index:++a});r.append(w[A],ot())}}}else if(8===r.nodeType)if(r.data===nt)s.push({type:2,index:a});else for(var O=-1;-1!==(O=r.data.indexOf(et,O+1));)s.push({type:7,index:a}),O+=et.length-1;a++}}return x(t,null,[{key:"createElement",value:function(t,e){var n=it.createElement("template");return n.innerHTML=t,n}}]),t}();function kt(t,e){var n,r,i,o,a,l=arguments.length>2&&void 0!==arguments[2]?arguments[2]:t,u=arguments.length>3?arguments[3]:void 0;if(e===yt)return e;var s=void 0!==u?null===(n=l._$Co)||void 0===n?void 0:n[u]:l._$Cl,c=at(e)?void 0:e._$litDirective$;return(null===(r=s)||void 0===r?void 0:r.constructor)!==c&&(null!==(i=s)&&void 0!==i&&null!==(o=i._$AO)&&void 0!==o&&o.call(i,!1),void 0===c?s=void 0:(s=new c(t))._$AT(t,l,u),void 0!==u?(null!==(a=l._$Co)&&void 0!==a?a:l._$Co=[])[u]=s:l._$Cl=s),void 0!==s&&(e=kt(t,s._$AS(t,e.values),s,u)),e}var At=function(){function t(e,n){S(this,t),this._$AV=[],this._$AN=void 0,this._$AD=e,this._$AM=n}return x(t,[{key:"parentNode",get:function(){return this._$AM.parentNode}},{key:"_$AU",get:function(){return this._$AM._$AU}},{key:"u",value:function(t){var e,n=this._$AD,r=n.el.content,i=n.parts,o=(null!==(e=null==t?void 0:t.creationScope)&&void 0!==e?e:it).importNode(r,!0);_t.currentNode=o;for(var a=_t.nextNode(),l=0,u=0,s=i[0];void 0!==s;){var c;if(l===s.index){var h=void 0;2===s.type?h=new Et(a,a.nextSibling,this,t):1===s.type?h=new s.ctor(a,s.name,s.strings,this,t):6===s.type&&(h=new jt(a,this,t)),this._$AV.push(h),s=i[++u]}l!==(null===(c=s)||void 0===c?void 0:c.index)&&(a=_t.nextNode(),l++)}return _t.currentNode=it,o}},{key:"p",value:function(t){var e,n=0,r=k(this._$AV);try{for(r.s();!(e=r.n()).done;){var i=e.value;void 0!==i&&(void 0!==i.strings?(i._$AI(t,i,n),n+=i.strings.length-2):i._$AI(t[n])),n++}}catch(t){r.e(t)}finally{r.f()}}}]),t}(),Et=function(){function t(e,n,r,i){var o;S(this,t),this.type=2,this._$AH=gt,this._$AN=void 0,this._$AA=e,this._$AB=n,this._$AM=r,this.options=i,this._$Cv=null===(o=null==i?void 0:i.isConnected)||void 0===o||o}return x(t,[{key:"_$AU",get:function(){var t,e;return null!==(t=null===(e=this._$AM)||void 0===e?void 0:e._$AU)&&void 0!==t?t:this._$Cv}},{key:"parentNode",get:function(){var t,e=this._$AA.parentNode,n=this._$AM;return void 0!==n&&11===(null===(t=e)||void 0===t?void 0:t.nodeType)&&(e=n.parentNode),e}},{key:"startNode",get:function(){return this._$AA}},{key:"endNode",get:function(){return this._$AB}},{key:"_$AI",value:function(t){t=kt(this,t,arguments.length>1&&void 0!==arguments[1]?arguments[1]:this),at(t)?t===gt||null==t||""===t?(this._$AH!==gt&&this._$AR(),this._$AH=gt):t!==this._$AH&&t!==yt&&this._(t):void 0!==t._$litType$?this.$(t):void 0!==t.nodeType?this.T(t):function(t){return lt(t)||"function"==typeof(null==t?void 0:t[Symbol.iterator])}(t)?this.k(t):this._(t)}},{key:"S",value:function(t){return this._$AA.parentNode.insertBefore(t,this._$AB)}},{key:"T",value:function(t){this._$AH!==t&&(this._$AR(),this._$AH=this.S(t))}},{key:"_",value:function(t){this._$AH!==gt&&at(this._$AH)?this._$AA.nextSibling.data=t:this.T(it.createTextNode(t)),this._$AH=t}},{key:"$",value:function(t){var e,n=t.values,r=t._$litType$,i="number"==typeof r?this._$AC(t):(void 0===r.el&&(r.el=wt.createElement(bt(r.h,r.h[0]),this.options)),r);if((null===(e=this._$AH)||void 0===e?void 0:e._$AD)===i)this._$AH.p(n);else{var o=new At(i,this),a=o.u(this.options);o.p(n),this.T(a),this._$AH=o}}},{key:"_$AC",value:function(t){var e=mt.get(t.strings);return void 0===e&&mt.set(t.strings,e=new wt(t)),e}},{key:"k",value:function(e){lt(this._$AH)||(this._$AH=[],this._$AR());var n,r,i=this._$AH,o=0,a=k(e);try{for(a.s();!(r=a.n()).done;){var l=r.value;o===i.length?i.push(n=new t(this.S(ot()),this.S(ot()),this,this.options)):n=i[o],n._$AI(l),o++}}catch(t){a.e(t)}finally{a.f()}o<i.length&&(this._$AR(n&&n._$AB.nextSibling,o),i.length=o)}},{key:"_$AR",value:function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:this._$AA.nextSibling,e=arguments.length>1?arguments[1]:void 0;for(null===(n=this._$AP)||void 0===n||n.call(this,!1,!0,e);t&&t!==this._$AB;){var n,r=t.nextSibling;t.remove(),t=r}}},{key:"setConnected",value:function(t){var e;void 0===this._$AM&&(this._$Cv=t,null===(e=this._$AP)||void 0===e||e.call(this,t))}}]),t}(),St=function(){function t(e,n,r,i,o){S(this,t),this.type=1,this._$AH=gt,this._$AN=void 0,this.element=e,this.name=n,this._$AM=i,this.options=o,r.length>2||""!==r[0]||""!==r[1]?(this._$AH=Array(r.length-1).fill(new String),this.strings=r):this._$AH=gt}return x(t,[{key:"tagName",get:function(){return this.element.tagName}},{key:"_$AU",get:function(){return this._$AM._$AU}},{key:"_$AI",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:this,n=arguments.length>2?arguments[2]:void 0,r=arguments.length>3?arguments[3]:void 0,i=this.strings,o=!1;if(void 0===i)t=kt(this,t,e,0),(o=!at(t)||t!==this._$AH&&t!==yt)&&(this._$AH=t);else{var a,l,u=t;for(t=i[0],a=0;a<i.length-1;a++){var s;(l=kt(this,u[n+a],e,a))===yt&&(l=this._$AH[a]),o||(o=!at(l)||l!==this._$AH[a]),l===gt?t=gt:t!==gt&&(t+=(null!==(s=l)&&void 0!==s?s:"")+i[a+1]),this._$AH[a]=l}}o&&!r&&this.j(t)}},{key:"j",value:function(t){t===gt?this.element.removeAttribute(this.name):this.element.setAttribute(this.name,null!=t?t:"")}}]),t}(),Ot=function(t){function e(){var t;return S(this,e),(t=y(this,e,arguments)).type=3,t}return m(e,St),x(e,[{key:"j",value:function(t){this.element[this.name]=t===gt?void 0:t}}]),e}(),xt=function(t){function e(){var t;return S(this,e),(t=y(this,e,arguments)).type=4,t}return m(e,St),x(e,[{key:"j",value:function(t){this.element.toggleAttribute(this.name,!!t&&t!==gt)}}]),e}(),Pt=function(t){function e(t,n,r,i,o){var a;return S(this,e),(a=y(this,e,[t,n,r,i,o])).type=5,a}return m(e,St),x(e,[{key:"_$AI",value:function(t){var e;if((t=null!==(e=kt(this,t,arguments.length>1&&void 0!==arguments[1]?arguments[1]:this,0))&&void 0!==e?e:gt)!==yt){var n=this._$AH,r=t===gt&&n!==gt||t.capture!==n.capture||t.once!==n.once||t.passive!==n.passive,i=t!==gt&&(n===gt||r);r&&this.element.removeEventListener(this.name,this,n),i&&this.element.addEventListener(this.name,this,t),this._$AH=t}}},{key:"handleEvent",value:function(t){var e,n;"function"==typeof this._$AH?this._$AH.call(null!==(e=null===(n=this.options)||void 0===n?void 0:n.host)&&void 0!==e?e:this.element,t):this._$AH.handleEvent(t)}}]),e}(),jt=function(){function t(e,n,r){S(this,t),this.element=e,this.type=6,this._$AN=void 0,this._$AM=n,this.options=r}return x(t,[{key:"_$AU",get:function(){return this._$AM._$AU}},{key:"_$AI",value:function(t){kt(this,t)}}]),t}(),Ct=Z.litHtmlPolyfillSupport;null!=Ct&&Ct(wt,Et),(null!==(r=Z.litHtmlVersions)&&void 0!==r?r:Z.litHtmlVersions=[]).push("3.1.2");var Nt=function(t){function e(){var t;return S(this,e),(t=y(this,e,arguments)).renderOptions={host:g(t)},t._$Do=void 0,t}return m(e,Y),x(e,[{key:"createRenderRoot",value:function(){var t,n,r=s(w(e.prototype),"createRenderRoot",this).call(this);return null!==(n=(t=this.renderOptions).renderBefore)&&void 0!==n||(t.renderBefore=r.firstChild),r}},{key:"update",value:function(t){var n=this.render();this.hasUpdated||(this.renderOptions.isConnected=this.isConnected),s(w(e.prototype),"update",this).call(this,t),this._$Do=function(t,e,n){var r,i=null!==(r=null==n?void 0:n.renderBefore)&&void 0!==r?r:e,o=i._$litPart$;if(void 0===o){var a,l=null!==(a=null==n?void 0:n.renderBefore)&&void 0!==a?a:null;i._$litPart$=o=new Et(e.insertBefore(ot(),l),l,void 0,null!=n?n:{})}return o._$AI(t),o}(n,this.renderRoot,this.renderOptions)}},{key:"connectedCallback",value:function(){var t;s(w(e.prototype),"connectedCallback",this).call(this),null===(t=this._$Do)||void 0===t||t.setConnected(!0)}},{key:"disconnectedCallback",value:function(){var t;s(w(e.prototype),"disconnectedCallback",this).call(this),null===(t=this._$Do)||void 0===t||t.setConnected(!1)}},{key:"render",value:function(){return yt}}]),e}();Nt._$litElement$=!0,Nt.finalized=!0,null===(i=globalThis.litElementHydrateSupport)||void 0===i||i.call(globalThis,{LitElement:Nt});var Lt=globalThis.litElementPolyfillSupport;null==Lt||Lt({LitElement:Nt}),(null!==(o=globalThis.litElementVersions)&&void 0!==o?o:globalThis.litElementVersions=[]).push("4.0.4");var Tt="code",Rt="pin_used",Ut="pin_synced_to_locks",It=["number_of_uses"],Mt={type:"divider"},Ht=["name","enabled","pin",Ut].concat(It,[Tt,Rt]),Dt="fold-entity-row.js";function zt(t){var e,n=arguments.length>1&&void 0!==arguments[1]?arguments[1]:"-",r="àáâäæãåāăąçćčđďèéêëēėęěğǵḧîïíīįìıİłḿñńǹňôöòóœøōõőṕŕřßśšşșťțûüùúūǘůűųẃẍÿýžźż·",i="aaaaaaaaaacccddeeeeeeeegghiiiiiiiilmnnnnoooooooooprrsssssttuuuuuuuuuwxyyzzz".concat(n),o=new RegExp(r.split("").join("|"),"g");return""===t?e="":(e=t.toString().toLowerCase().replace(o,(function(t){return i.charAt(r.indexOf(t))})).replace(/(\d),(?=\d)/g,"$1").replace(/[^a-z0-9]+/g,n).replace(new RegExp("(".concat(n,")\\1+"),"g"),"$1").replace(new RegExp("^".concat(n,"+")),"").replace(new RegExp("".concat(n,"+$")),""),""===e&&(e="unknown")),e}var Bt=function(t){function e(){return S(this,e),y(this,e,arguments)}var n;return m(e,Y),x(e,null,[{key:"generate",value:(n=p(d().mark((function t(e,n){var r,i;return d().wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,n.callWS({domain:"lock_code_manager",type:"config_entries/get"});case 2:if(0!==(r=t.sent).length){t.next=5;break}return t.abrupt("return",{title:"Lock Code Manager",views:[{badges:[],cards:[{content:"# No Lock Code Manager configurations found!",type:"markdown"}],title:"Lock Code Manager"}]});case 5:return t.next=7,Promise.all(r.map((function(t){var n;return{path:zt(t.title),strategy:{config_entry_id:t.entry_id,include_code_slot_sensors:null!==(n=e.include_code_slot_sensors)&&void 0!==n&&n,type:"custom:lock-code-manager"},title:t.title}})));case 7:return 1===(i=t.sent).length&&i.push({title:"​"}),t.abrupt("return",{title:"Lock Code Manager",views:i});case 10:case"end":return t.stop()}}),t)}))),function(t,e){return n.apply(this,arguments)})}]),e}();function Wt(t,e,n,r,i){return Vt.apply(this,arguments)}function Vt(){return Vt=p(d().mark((function t(e,n,r,i,o){var a,l,u,s,c,v,p,y,g,m;return d().wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,Promise.all([e.callWS({config_entry_id:n,type:"lock_code_manager/get_slot_calendar_data"}),e.callWS({type:"lovelace/resources"})]);case 2:return a=t.sent,l=f(a,2),u=l[0],s=l[1],c=i.map((function(t){return Ft(t)})).sort(Gt),v=Object.keys(u.slots).map((function(t){return parseInt(t,10)})),p=v.map((function(t){return Kt(e,t,c,u)})),y=[].concat(h(u.locks.sort((function(t,e){return t.localeCompare(e)}))),h(c.filter((function(t){return"pin_synced_to_locks"===t.key})).map((function(t){return{entity:t.entity_id,name:(t.name?t.name:t.original_name).replace("PIN synced to locks","synced").replace("Code slot","Slot"),type:"state-label"}})))),g=s.filter((function(t){return t.url.includes(Dt)})).length>0,m=p.map((function(t){return Jt(t,g,o)})),t.abrupt("return",{badges:y,cards:m,panel:!1,path:zt(r),title:r});case 13:case"end":return t.stop()}}),t)}))),Vt.apply(this,arguments)}function Gt(t,e){return t.slotNum<e.slotNum?-1:t.slotNum>e.slotNum?1:Ht.indexOf(t.key)<Ht.indexOf(e.key)?-1:Ht.indexOf(t.key)>Ht.indexOf(e.key)?1:t.key===e.key&&[Rt,Tt].includes(t.key)&&t.lockEntityId<e.lockEntityId?-1:1}function Ft(t){var e=t.unique_id.split("|");return l(l({},t),{},{key:e[2],lockEntityId:e[3],slotNum:parseInt(e[1],10)})}function qt(t){return t.map((function(t){return{entity:t}}))}function Jt(t,e,n){return{cards:[{content:"## Code Slot ".concat(t.slotNum),type:"markdown"},{entities:[].concat(h(qt(t.mainEntityIds)),[Mt,{entity:t.pinShouldBeEnabledEntity.entity_id,name:"PIN synced to locks"},{entity:t.codeEventEntityId,name:"PIN Last Used"}],h(Yt(t.conditionEntityIds,"Conditions",e)),h(n?Yt(t.codeSensorEntityIds,"Code Slot Sensors",e):[])),show_header_toggle:!1,type:"entities"}],type:"vertical-stack"}}function Kt(t,e,n,r){var i,o=[],a=[],l=[];n.filter((function(t){return t.slotNum===e})).forEach((function(t){t.key===Tt?l.push(t.entity_id):t.key===Rt?i=t.entity_id:It.includes(t.key)?a.push(t.entity_id):t.key!==Ut&&o.push(t.entity_id)}));var u=n.find((function(t){return t.slotNum===e&&t.key===Ut})),s=r.slots[e];return s&&a.unshift(s),{codeEventEntityId:i,codeSensorEntityIds:l,conditionEntityIds:a,mainEntityIds:o,pinShouldBeEnabledEntity:u,slotNum:e}}function Yt(t,e,n){if(0===t.length)return[];var r=qt(t);return n?[Mt,{entities:r,head:{label:e,type:"section"},type:"custom:fold-entity-row"}]:[{label:e,type:"section"}].concat(h(r))}var Zt=function(t){function e(){return S(this,e),y(this,e,arguments)}var n;return m(e,Y),x(e,null,[{key:"generate",value:(n=p(d().mark((function t(e,n){var r,i,o,a,l,u,s,c,h;return d().wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(r=e.config_entry_id,i=e.config_entry_title,!(void 0===r&&void 0===i||void 0!==r&&void 0!==i)){t.next=3;break}return t.abrupt("return",{badges:[],cards:[{content:"## ERROR: Either `config_entry_title` or `config_entry_id` must be provided in the view config, but not both!",type:"markdown"}],title:"Lock Code Manager"});case 3:return t.prev=3,t.next=6,n.callWS({config_entry_id:r,config_entry_title:i,type:"lock_code_manager/get_config_entry_entities"});case 6:return a=t.sent,l=f(a,3),u=l[0],s=l[1],c=l[2],t.abrupt("return",Wt(n,u,s,c,null!==(o=e.include_code_slot_sensors)&&void 0!==o&&o));case 14:return t.prev=14,t.t0=t.catch(3),h=void 0!==r?"with ID `".concat(r,"`"):"called `".concat(i,"`"),t.abrupt("return",{badges:[],cards:[{content:"## ERROR: No Lock Code Manager configuration ".concat(h," found!"),type:"markdown"}],title:"Lock Code Manager"});case 18:case"end":return t.stop()}}),t,null,[[3,14]])}))),function(t,e){return n.apply(this,arguments)})}]),e}();customElements.define("ll-strategy-dashboard-lock-code-manager",Bt),customElements.define("ll-strategy-view-lock-code-manager",Zt);
