+++ js/dashboard-strategy.ts
@@ -1,8 +1,9 @@
 import { ReactiveElement } from 'lit';
 
 import { DEFAULT_INCLUDE_CODE_SLOT_SENSORS } from './const';
+import { generateView } from './generate-view';
 import { HomeAssistant } from './ha_type_stubs';
-import { generateView, slugify } from './helpers';
+import { slugify } from './slugify';
 import { GetConfigEntriesResponse, LockCodeManagerDashboardStrategyConfig } from './types';
 
 export class LockCodeManagerDashboardStrategy extends ReactiveElement {
