+++ js/view-strategy.ts
@@ -1,8 +1,8 @@
 import { ReactiveElement } from 'lit';
 
 import { DEFAULT_INCLUDE_CODE_SLOT_SENSORS } from './const';
+import { generateView } from './generate-view';
 import { HomeAssistant } from './ha_type_stubs';
-import { generateView } from './helpers';
 import { LockCodeManagerEntitiesResponse, LockCodeManagerViewStrategyConfig } from './types';
 
 export class LockCodeManagerViewStrategy extends ReactiveElement {
